Added a function to check that the sensor is available.
Noted that the AS3935 class can be instatiated on a board with nothing connected to the SPI bus.
Added a _spi_device_exists method to call from __init__ after instantiating the sensor. If nothing is connected, SPI
reads return 0x00. The indoor register only returns none zero values. If it returns 0x00, raise an OSError.